+++ lab7/iterable/CMakeLists.txt
@@ -1,8 +1,8 @@
 
 project(iterable)
 
-set(HEDEAR_FILES Iterable.h Enumerate.h Product.h Zipper.h)
-set(SOURCE_FILES Iterable.cpp Enumerate.cpp Product.cpp Zipper.cpp)
+set(HEDEAR_FILES Iterable.h )
+set(SOURCE_FILES Iterable.cpp )
 
 add_library(libiterable STATIC ${SOURCE_FILES})
 target_include_directories(libiterable PUBLIC .)
